refactor: move view to CreeDictionary.views
